requirements.txt
Django~=5.0.6
openpyxl~=3.1.5
Django-import-export~=3.2.0